requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
mysql-connector-python==9.1.0
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
